+++ CHANGELOG.md
@@ -14,6 +14,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 - upcloud_network: detect if resource was deleted outside of Terraform
 - upcloud_network_peering: detect if resource was deleted outside of Terraform
 - upcloud_floating_ip_address: replace floating IP address, if `family` or `zone` have changes.
+- provider: do not replace zero value with default when configuring plugin framework provider. For example, `request_timeout_sec = 0` will now disable request timeout also for resources migrated to plugin framework (e.g. `upcloud_network`).
 
 ## [5.5.0] - 2024-06-04